+++ rust/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 name = "rust"
 version = "0.1.0"
-edition = "2021"
+edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [profile.release]
-#panic = "abort"
+panic = "abort"
 
 [dependencies]
+libc = "0.2"
